+++ server/src/main/kotlin/common/Video.kt
@@ -1,3 +1,3 @@
 package common
 
-data class Video(val id: Long, val title: String, val videoUrl: String)
+data class Video(val id: Int, val title: String, val videoUrl: String)
